[ContentSlider] Extends component from TemplatesComponent to define the items by html templates
