+++ demos/content-slider/src/ts/components/content-slider-example/content-slider-example.component.html
@@ -1,1 +1,58 @@
-<content-slider></content-slider>
+<content-slider>
+  <template image-src="../../../images/atimetostand.jpg">
+    <p class="lead">29. June 2020 19.00h - 22.00h</p>
+    <h2 class="display-3">A Time to Stand & Backlash</h2>
+    <p class="lead">
+        Chocolate sugar plum candy oat cake. Carrot cake croissant ice cream gummies marshmallow sweet roll pudding icing marzipan. Muffin tootsie roll dragée. Oat cake sweet roll <a href="https://www.facebook.com/atimetostand/" target="_blank">chocolate cake</a>.
+    </p>
+  </template>
+  <template image-src="../../../images/raum27.jpg">
+    <p class="lead">15. July 2020 19.00h - 22.00h</p>
+    <h2 class="display-3">She Danced Slowly</h2>
+    <p class="lead">
+        Tootsie roll topping cake bonbon gummi bears icing. Pastry ice <a href="https://www.facebook.com/shedancedslowly/" target="_blank">cream cupcake apple pie</a>. Lemon drops macaroon bear claw bear claw dragée halvah. Sesame snaps bonbon liquorice tiramisu sweet roll caramels donut. Cookie danish macaroon dessert dessert marshmallow dessert jelly beans. 
+    </p>
+  </template>
+  <template image-src="../../../images/shamotte.jpg">
+    <p class="lead">01. April 2020 19.00h - 22.00h</p>
+    <h2 class="display-3">Arr gunwalls hardtack cackle</h2>
+    <p class="lead">
+        Warp Plate Fleet lookout hail-shot walk the plank grog blossom sheet main sheet bilge water run a rig. Strike colors hardtack jolly boat long clothes plunder carouser fire ship square-rigged ahoy belaying pin. Marooned Davy Jones' Locker gibbet sheet yardarm Pirate Round boom Nelsons folly tack lanyard. 
+    </p>
+  </template>
+  <template image-src="../../../images/shoreline.jpg">
+    <p class="lead">22. March 2020 18.00h - 21.00h</p>
+    <h2 class="display-3">Warp Plate lookout hail-shot</h2>
+    <p class="lead">
+        Sugar plum marzipan gummi bears cupcake tiramisu candy halvah. Caramels danish jujubes bear claw chocolate dessert. Carrot cake jelly dessert jelly-o. Gummies halvah sugar plum carrot cake tiramisu cake. Gummies fruitcake lollipop tootsie roll lemon drops donut apple pie.
+    </p>
+  </template>
+  <template image-src="../../../images/jerakeen.jpg">
+    <p class="lead">30. September 2020 19.00h - 22.00h</p>
+    <h2 class="display-3">Code of conduct Sea Legs hardtack</h2>
+    <p class="lead">
+        Marshmallow chocolate bar topping. Bear claw cupcake toffee liquorice. Powder bonbon croissant. Pastry cotton candy marzipan lollipop jelly candy canes lemon drops cotton candy dessert. Toffee candy sweet roll halvah apple pie pudding chocolate bar dragée pie. Marshmallow chocolate bar sesame snaps dessert caramels sesame snaps jelly-o. Fruitcake apple pie soufflé marshmallow bear claw cheesecake. Powder gingerbread oat cake bonbon jujubes soufflé. Candy candy oat cake cookie cake toffee. Tart powder caramels liquorice. Bonbon gingerbread marzipan liquorice donut jelly caramels jelly. Lollipop toffee tiramisu jelly-o marshmallow soufflé.
+    </p>
+  </template>
+  <template image-src="https://picsum.photos/200/300">
+    <p class="lead">03. August 2020 23.59h - 08.00h</p>
+    <h2 class="display-3">Davy Jones' Locker</h2>
+    <p class="lead">
+        Tootsie roll topping cake bonbon gummi bears icing. Pastry ice <a href="https://www.facebook.com/shedancedslowly/" target="_blank">cream cupcake apple pie</a>. Lemon drops macaroon bear claw bear claw dragée halvah. Sesame snaps bonbon liquorice tiramisu sweet roll caramels donut. Cookie danish macaroon dessert dessert marshmallow dessert jelly beans. 
+    </p>
+  </template>
+  <template image-src="https://picsum.photos/1200/1800">
+    <p class="lead">03. May 2020 19.00h - 22.00h</p>
+    <h2 class="display-3">Coxswain coffer black</h2>
+    <p class="lead">
+        Chocolate bar brownie soufflé cake fruitcake sweet roll. Lollipop topping dragée caramels topping candy canes. Biscuit pie chocolate powder liquorice topping sweet roll dessert danish. Gummies oat cake candy canes gummies danish muffin. Bear claw tart cake caramels tiramisu. 
+    </p>
+  </template>
+  <template image-src="https://picsum.photos/900/600">
+    <p class="lead">13. November 2020 23.00h - 05.00h</p>
+    <h2 class="display-3">Marooned Davy Jones' Locker</h2>
+    <p class="lead">
+        Code of conduct Sea Legs hardtack to go on account Sail ho hang the jib bilge water Plate Fleet strike colors heave down. Gold Road warp rum rutters spirits overhaul topmast Nelsons folly jib hail-shot. Coxswain coffer black spot six pounders gabion bilge water lass haul wind mutiny square-rigged. 
+    </p>
+  </template>
+</content-slider>
