+++ demos/content-slider/src/ts/components/content-slider-example/content-slider-example.component.scss
@@ -1,26 +1,5 @@
-video-example {
-  .mobile {
-    display: none;
-  }
-  .desktop {
-    display: block;
-  }
+content-slider {
 
-  @media (max-width: 1200px) {
-    .mobile {
-      display: block;
-    }
-    .desktop {
-      display: none;
-    }
-  }
-
-  rv-video, video {
-    width: 100vw;
-    height: 100vh;
-    max-width: 100vw;
-    max-height: 100vh;
-  }
 }
