+++ packages/core/src/components/index.ts
@@ -1,1 +1,2 @@
+export { TemplatesComponent } from "./templates/templates.component";
 export { VideoComponent } from "./video/video.component";
